refactor to components directory
