+++ src/components/stopwatch/index.ts
@@ -1,0 +1,1 @@
+export {default as StopWatch} from './StopWatch';
